# services/gateway-api/openapi.yaml
openapi: "3.1.0"
info:
  title: Opencomplai Gateway API
  version: 0.9.0
  description: |
    REST API gateway for the Opencomplai compliance platform.
    All routes proxy to the appropriate Python service.
    This spec is the source of truth for the API contract.

    Manifests can name several frameworks (compliance_targets). Risk
    classification, dossier generation and the risk engine's public
    /v1/checker/* routes (called by the docs widget, not proxied here) are
    EU AI Act only.

    Authentication: every /v1 route requires credentials, either an
    x-api-key header (API-key mode) or an Authorization bearer JWT (OIDC
    mode). A deployment uses one of the two. Only GET /health is open. The
    gateway refuses to start unconfigured unless authentication is explicitly
    disabled for local development (OPENCOMPLAI_AUTH_DISABLED=1).
servers:
  - url: http://localhost:8080
    description: Local Docker Compose deployment
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /v1/hitl/overrides:
    post:
      operationId: submitOverride
      summary: Submit a HITL override action with mandatory rationale (idempotent by idempotency_key)
      description: >-
        The decision is written to the evidence ledger before it is accepted;
        if the ledger is unavailable the request fails with 503 and nothing is
        recorded. Repeating a request with the same idempotency_key (derived
        from the payload when omitted) replays the original response. The
        gateway replaces actor_id with the authenticated caller's identity
        (the JWT sub claim, or api-key-caller in API-key mode); it keeps the
        value you send only when authentication is disabled or the token has
        no sub claim.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OverrideRequest"
      responses:
        "201":
          description: Override submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverrideResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: idempotency_key was already used with a different payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamError"
        "422":
          description: >-
            Missing or empty case_id, actor_id or rationale, or decision not
            approved/rejected (gateway envelope); a whitespace-only rationale
            is rejected by the risk engine (upstream error shape)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorEnvelope"
                  - $ref: "#/components/schemas/UpstreamError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/LedgerOrDependencyUnavailable"
  /v1/hitl/overrides/{override_id}/second-approval:
    post:
      operationId: secondApproveOverride
      summary: Complete a dual-control override with a second approver
      description: >-
        Completes an override submitted with requires_dual_approval. The
        second approver must differ from the first and must send the first
        approval's rationale_hash plus their own non-empty rationale. The
        override_second_approved event is written to the evidence ledger
        before the request is accepted. The gateway replaces actor_id with
        the authenticated caller's identity; in API-key mode that is always
        api-key-caller, so two distinct approvers need JWT/OIDC.
      parameters:
        - name: override_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SecondApprovalRequest"
      responses:
        "201":
          description: Override completed (status accepted or rejected)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverrideResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Same approver as the first approval (POLICY_DENIED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamError"
        "404":
          description: No pending dual-control override with this id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamError"
        "409":
          description: ALREADY_COMPLETED or RATIONALE_HASH_MISMATCH
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamError"
        "422":
          description: Missing or empty field (gateway envelope), or whitespace-only rationale (upstream error shape)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorEnvelope"
                  - $ref: "#/components/schemas/UpstreamError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/LedgerOrDependencyUnavailable"
  /v1/hitl/queue:
    get:
      operationId: listReviewQueue
      summary: List items in the human-review queue (idempotent)
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum: [queued, assigned, decided, expired]
          description: Return only items in this state
        - name: assigned_to
          in: query
          required: false
          schema:
            type: string
          description: Return only items assigned to this reviewer
      responses:
        "200":
          description: Matching review items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewQueueListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
  /v1/hitl/queue/{id}:
    get:
      operationId: getReviewQueueItem
      summary: Get one review item with its redacted reviewer context (idempotent)
      parameters:
        - $ref: "#/components/parameters/ReviewId"
      responses:
        "200":
          description: The review item and its context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewQueueDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Review item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
  /v1/hitl/queue/{id}/assign:
    post:
      operationId: assignReviewQueueItem
      summary: Assign a review item to a reviewer
      description: >-
        Sets the item to assigned for the named reviewer. The gateway replaces
        reviewer_id with the authenticated caller's identity (the JWT sub
        claim, or api-key-caller in API-key mode); it keeps the value you send
        only when authentication is disabled or the token has no sub claim.
      parameters:
        - $ref: "#/components/parameters/ReviewId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewAssignRequest"
      responses:
        "200":
          description: Updated review item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewAssignResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Review item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamError"
        "422":
          description: Missing or empty reviewer_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
  /v1/hitl/queue/{id}/decide:
    post:
      operationId: decideReviewQueueItem
      summary: Record the decision on a review item with mandatory rationale
      description: >-
        Submits a HITL override with the review id as the case and marks the
        item decided. When no idempotency_key is given, one is derived from
        the review id, so a second decision on the same item replays the first
        unless its payload differs (409). The gateway replaces actor_id with
        the authenticated caller's identity (the JWT sub claim, or
        api-key-caller in API-key mode); it keeps the value you send only when
        authentication is disabled or the token has no sub claim.
      parameters:
        - $ref: "#/components/parameters/ReviewId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecideRequest"
      responses:
        "201":
          description: Decision recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewDecideResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Review item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamError"
        "409":
          description: idempotency_key was already used with a different payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamError"
        "422":
          description: >-
            Missing or empty actor_id or rationale, or decision not
            approved/rejected (gateway envelope); a whitespace-only rationale
            is rejected by the risk engine (upstream error shape)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorEnvelope"
                  - $ref: "#/components/schemas/UpstreamError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/LedgerOrDependencyUnavailable"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: >-
        API-key mode, for self-hosted single-operator deployments: the value
        of OPENCOMPLAI_API_KEY. A shared key authenticates the deployment as a
        whole, as the tenant named by OPENCOMPLAI_API_KEY_TENANT_ID (default
        oss-default). Ignored when OIDC_JWKS_URI is set. See
        docs/src/deployment/authentication.md.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        OIDC mode, for multi-user deployments (OIDC_JWKS_URI set): an RS256
        JWT verified against the identity provider's JWKS, with iss, aud, exp
        and nbf enforced. The token must also carry the tenant claim named by
        OIDC_TENANT_CLAIM. See docs/src/deployment/authentication.md.
  headers:
...
    RetryAfter:
      description: Seconds to wait before retrying; sent only with the rate limiter's own 429
      schema:
        type: integer
  parameters:
    ReviewId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Review id as returned by GET /v1/hitl/queue (rev_sha256:<hex>)
...
  responses:
    Unauthorized:
      description: >-
        Credentials missing or invalid (POLICY_DENIED). The message says
        which: "Invalid or missing API key." in API-key mode; "Missing Bearer
        token.", "Invalid or expired JWT." or "Token missing tenant claim." in
        OIDC mode.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimited:
      description: >-
        Too many requests. After 10 failed authentication attempts from one
        client within 60 seconds (the defaults) further failures return 429
        RATE_LIMITED in the gateway envelope instead of 401. Separately, more
        than 300 requests of any kind from one client within 60 seconds (the
        default; the Docker Compose stack raises it to 1000) return the rate
        limiter's own body, {statusCode, error, message}, with a Retry-After
        header. Both limits are configurable.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/ErrorEnvelope"
              - $ref: "#/components/schemas/RequestRateLimitError"
...
    DependencyUnavailable:
      description: An upstream service could not be reached
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DependencyUnavailableError"
    LedgerOrDependencyUnavailable:
      description: >-
        The evidence ledger is unavailable and the decision was not recorded
        (AUDIT_UNAVAILABLE, upstream error shape), or an upstream service
        could not be reached (DEPENDENCY_UNAVAILABLE)
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/DependencyUnavailableError"
              - $ref: "#/components/schemas/UpstreamError"
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error_code, message, category, retryable, correlation_id]
      properties:
        error_code:
          type: string
        message:
          type: string
        category:
          type: string
          enum: [client, policy, system, dependency]
        retryable:
          type: boolean
        correlation_id:
          type: string
        remediation_hint:
          type: string
    RequestRateLimitError:
      type: object
      required: [statusCode, error, message]
      description: >-
        Body the gateway's request rate limiter sends when a client exceeds
        the global limit. It does not use the ErrorEnvelope shape.
      properties:
        statusCode:
          type: integer
          enum: [429]
        error:
          type: string
          enum: [Too Many Requests]
        message:
          type: string
          description: Rate limit exceeded, retry in <duration>
...
    UpstreamError:
      type: object
      required: [detail]
      description: >-
        Error body from an internal service, forwarded unchanged by the
        gateway. The services are FastAPI apps, so the body is
        {"detail": ...}; detail is a string, an object (for example
        {error_code, message, category, retryable}) or, for request validation
        failures, a list of error objects.
      properties:
        detail: {}
    DependencyUnavailableError:
      type: object
      required: [error_code, message, retryable]
      description: >-
        Returned with 503 when an upstream service cannot be reached. The
        gateway adds category and correlation_id; the egress proxy's variant
        carries only error_code, message and retryable.
      properties:
        error_code:
          type: string
          enum: [DEPENDENCY_UNAVAILABLE]
        message:
          type: string
        category:
          type: string
          enum: [dependency]
        retryable:
          type: boolean
        correlation_id:
          type: string
...
    SecondApprovalRequest:
      type: object
      required: [actor_id, rationale, rationale_hash]
      properties:
        actor_id:
          type: string
          minLength: 1
          description: Replaced by the authenticated caller's identity
        rationale:
          type: string
          minLength: 1
          description: The second approver's own rationale; only its hash is stored
        rationale_hash:
          type: string
          minLength: 1
          description: The first approval's rationale_hash
        decision:
          type: string
          enum: [approved, rejected]
          default: approved
    OverrideRequest:
      type: object
      required: [case_id, actor_id, rationale, decision]
      properties:
        case_id:
          type: string
          minLength: 1
        actor_id:
          type: string
          minLength: 1
          description: >-
            Replaced by the authenticated caller's identity (the JWT sub
            claim, or api-key-caller in API-key mode); kept as sent only when
            authentication is disabled or the token has no sub claim
        rationale:
          type: string
          minLength: 1
        decision:
          type: string
          enum: [approved, rejected]
        requires_dual_approval:
          type: boolean
          description: >-
            When true the response status is pending_second_approval until a
            second approver completes it
        idempotency_key:
          type: string
          description: >-
            Same key with the same payload returns the original response;
            the same key with a different payload is rejected with 409.
            Derived from the payload when omitted.
    OverrideResponse:
      type: object
      required: [override_id, rationale_hash, status]
      properties:
        override_id:
          type: string
        rationale_hash:
          type: string
          description: sha256:<hex> of the rationale; the rationale text itself is not stored
        status:
          type: string
          enum: [accepted, pending_second_approval]
          description: >-
            Requests through the gateway are always accepted; it does not
            forward requires_dual_approval, so pending_second_approval is not
            reachable here.
        vault_event_id:
          type: [string, "null"]
          description: Id of the ledger event that recorded the override
    ReviewItem:
      type: object
      required:
        - review_id
        - tenant_id
        - system_id
        - commit_ref
        - reason
        - state
        - payload_ref
        - context_ref
        - idempotency_key
        - created_at
      properties:
        review_id:
          type: string
        tenant_id:
          type: string
        system_id:
          type: string
        commit_ref:
          type: string
        reason:
          type: string
          enum:
            - low_confidence
            - evaluator_fail
            - modification_trap
            - policy_block
            - manual
            - manifest_discrepancy
            - evidence_stale
            - manifest_change
        state:
          type: string
          enum: [queued, assigned, decided, expired]
        payload_ref:
          type: string
        context_ref:
          type: string
        reviewer_group:
          type: [string, "null"]
        assigned_to:
          type: [string, "null"]
        idempotency_key:
          type: string
        created_at:
          type: string
        expires_at:
          type: [string, "null"]
        decided_at:
          type: [string, "null"]
        linked_override_id:
          type: [string, "null"]
          description: Override recorded by the decision, once decided
    RedactedReviewContext:
      type: object
      description: Reviewer-safe context; carries no raw prompts, outputs or personal data.
      required:
        - review_id
        - reason
        - detector_ids
        - masked_excerpts
        - aggregate_counts
        - evidence_hashes
      properties:
        review_id:
          type: string
        reason:
          type: string
        detector_ids:
          type: array
          items:
            type: string
        masked_excerpts:
          type: array
          items:
            type: string
        aggregate_counts:
          type: object
          additionalProperties:
            type: integer
        evidence_hashes:
          type: array
          items:
            type: string
    ReviewQueueListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReviewItem"
    ReviewQueueDetailResponse:
      type: object
      required: [item, context]
      properties:
        item:
          $ref: "#/components/schemas/ReviewItem"
        context:
          oneOf:
            - $ref: "#/components/schemas/RedactedReviewContext"
            - type: "null"
    ReviewAssignRequest:
      type: object
      required: [reviewer_id]
      properties:
        reviewer_id:
          type: string
          minLength: 1
          description: >-
            Replaced by the authenticated caller's identity (the JWT sub
            claim, or api-key-caller in API-key mode); kept as sent only when
            authentication is disabled or the token has no sub claim
    ReviewAssignResponse:
      type: object
      required: [item]
      properties:
        item:
          $ref: "#/components/schemas/ReviewItem"
    ReviewDecideRequest:
      type: object
      required: [actor_id, rationale, decision]
      properties:
        actor_id:
          type: string
          minLength: 1
          description: >-
            Replaced by the authenticated caller's identity (the JWT sub
            claim, or api-key-caller in API-key mode); kept as sent only when
            authentication is disabled or the token has no sub claim
        rationale:
          type: string
          minLength: 1
          description: Mandatory; must contain non-whitespace characters
        decision:
          type: string
          enum: [approved, rejected]
        idempotency_key:
          type: string
          description: Defaults to decide-<review id>
    ReviewDecideResponse:
      type: object
      required: [item, override]
      properties:
        item:
          $ref: "#/components/schemas/ReviewItem"
        override:
          $ref: "#/components/schemas/OverrideResponse"
...